+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add CI tests for native modules
 - on Windows, build the corresponding arch. with the node under use
 - allow executing files within the enclosed package
+- upgrade libsquash to https://github.com/pmq20/libsquash/commit/f5b1a8d
 
 ## v0.9.6