+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - allow reusing the package itself as an Node.js interpreter
 - on Windows, build the corresponding arch. with the node under use
 - remove the `ENCLOSE_IO_ALWAYS_USE_ORIGINAL_NODE` hack
-- add auto-update feature via --auto-update-url and --auto-update-base-url
+- add auto-update feature via --auto-update-url and --auto-update-base
 - upgrade libsquash to https://github.com/pmq20/libsquash/commit/f5b1a8d
 
 ## v0.9.6